Reorganize UI tests into __tests__ folders
Moves several frontend spec files into colocated `__tests__` directories across setup, utils, vite, and shared store modules. Relative imports in the moved tests were updated (`.` to `..`/`../...`) so the test targets continue resolving correctly after the path changes.
